requirements.txt
selenium==4.15.2
schedule==1.2.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
pytest==7.4.3
pyyaml==6.0.1